Abstract out UploadMedia
