+++ web/src/components/AddMedia/UploadMedia.css
@@ -1,0 +1,14 @@
+.upload-media-drop-zone {
+    display: block;
+    background-color: #f0f0f0;
+    width: 100%;
+    padding: 40px;
+    text-align: center;
+    color: #838383;
+    font-size: x-large;
+    border-radius: 4px;
+}
+
+.upload-media-drop-zone:hover {
+    cursor: pointer;
+}
